requirements.txt
google-cloud-bigtable==0.31.1
google-cloud-core==0.28.1
connexion[swagger-ui]
Flask